+++ packages/stream_chat/CHANGELOG.md
@@ -19,6 +19,7 @@
 🐞 Fixed
 
 - [[#659]](https://github.com/GetStream/stream-chat-flutter/issues/659) Fixed unread count not updating correctly.
+- Fix `Filter.empty()` json encoding.
 
 ## 2.2.1